pyproject.toml
[tool.poetry]
name = "user-gallery"
version = "0.1.0"
description = ""
authors = ["Nikita Kostolomov"]
readme = "README.md"
packages = [{include = "user_gallery"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2-binary = "^2.9.5"
boto3 = "^1.26.73"
pillow = "^9.4.0"
django-storages = "^1.13.2"
drf-yasg = "^1.21.5"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.0.4"
commitizen = "^2.40.0"
django-environ = "^0.9.0"
pytest = "^7.2.2"
pytest-django = "^4.5.2"
requests = "^2.28.2"

[tool.black]
line-length = 79
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
